+++ src/index.js
@@ -1,2 +1,2 @@
 export { create } from './core';
-export * as sagaEffects from './saga-effects';
+export * from './saga-effects';