+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.4.0 - 2019-04-13
+
+### Udpated
+
+-   `@ackee/petrus@3.7.0`
+
+### Added
+
+-   Dispatch `verifyAccessTokenAvailability` action by `@ackee/petrus` on `401` response code status.
+
 ## 2.3.6 - 2019-02-25
 
 ### Updated